% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{clean_coincap_candles}
\alias{clean_coincap_candles}
\title{Helper function for processing CoinCap API OHLC data}
\usage{
clean_coincap_candles(df)
}
\arguments{
\item{df}{data frame returned by \code{coincap_ohlc_helper()}}
}
\value{
data frame
}
\description{
Helper function for processing CoinCap API OHLC data
}
