% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{coincap_granularities}
\alias{coincap_granularities}
\title{Helper function for CoinCap API OHLC interval granularities}
\usage{
coincap_granularities()
}
\value{
named list of granularities with useful parameters
}
\description{
Helper function for CoinCap API OHLC interval granularities
}
