% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{coincap_granularity}
\alias{coincap_granularity}
\title{Helper function for CoinCap API granularity parameters}
\usage{
coincap_granularity(granularity = "one_day", type = "interval")
}
\arguments{
\item{granularity}{character spelling out (number)_(granularity abbreviation)}

\item{type}{parameter to return}
}
\value{
character/numeric depending on parameter
}
\description{
Helper function for CoinCap API granularity parameters
}
