% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{coincap_markets}
\alias{coincap_markets}
\title{Query for available markets from the CoinCap API}
\usage{
coincap_markets(
  exchange_id,
  base_symbol,
  quote_symbol,
  base_id,
  quote_id,
  asset_symbol,
  asset_id,
  limit = 2000,
  offset = 1
)
}
\arguments{
\item{exchange_id}{search by exchange id}

\item{base_symbol}{returns all containing the base symbol}

\item{quote_symbol}{returns all containing the quote symbol}

\item{base_id}{returns all containing the base id}

\item{quote_id}{returns all containing the quote id}

\item{asset_symbol}{returns all assets containing symbol (base and quote)}

\item{asset_id}{returns all assets containing id (base and quote)}

\item{limit}{max limit of 2000}

\item{offset}{offset}
}
\value{
data frame
}
\description{
Query for available markets from the CoinCap API
}
