% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{coincap_ohlc}
\alias{coincap_ohlc}
\title{Query for available markets from the CoinCap API}
\usage{
coincap_ohlc(
  base_id = "bitcoin",
  quote_id = "tether",
  granularity = "one_day",
  exchange = "binance",
  start_tm = Sys.time() - (coincap_ohlc_iter_adj(granularity)/1000L),
  end_tm = Sys.time()
)
}
\arguments{
\item{base_id}{base id}

\item{quote_id}{quote id}

\item{exchange}{exchange id}

\item{interval}{candle interval}

\item{start}{UNIX time in milliseconds, omitting will return the most recent candles (ex: \code{1528410925604})}

\item{end}{UNIX time in milliseconds, omitting will return the most recent candles (ex: \code{1528411045604})}
}
\value{
data frame
}
\description{
Query for available markets from the CoinCap API
}
