% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{coincap_ohlc_helper}
\alias{coincap_ohlc_helper}
\title{Helper function for CoinCap API OHLC data}
\usage{
coincap_ohlc_helper(q_ls)
}
\arguments{
\item{q_ls}{named list of query key/name value pairs}
}
\value{
data frame
}
\description{
Helper function for CoinCap API OHLC data
}
