% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{coincap_ohlc_iter_adj}
\alias{coincap_ohlc_iter_adj}
\title{Helper function for CoinCap API granularity interval max values}
\usage{
coincap_ohlc_iter_adj(granularity)
}
\arguments{
\item{granularity}{character value returned by \code{coincap_granularity()} using \code{type = "interval"}}
}
\value{
character/numeric depending on parameter
}
\description{
Helper function for CoinCap API granularity interval max values
}
