% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{coincap_query_helper}
\alias{coincap_query_helper}
\title{Helper function for CoinCap API URL with query}
\usage{
coincap_query_helper(endpoint, q_ls)
}
\arguments{
\item{endpoint}{character}

\item{q_ls}{named list of query key/name value pairs}
}
\value{
character API URL
}
\description{
Helper function for CoinCap API URL with query
}
