% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{symbol_nm_map}
\alias{symbol_nm_map}
\alias{coincap_symbol_to_nm}
\alias{coincap_nm_to_symbol}
\title{Map between coin symbols and names}
\usage{
coincap_symbol_to_nm(x)

coincap_nm_to_symbol(x)
}
\arguments{
\item{x}{character symbol/name}
}
\value{
character symbol/name
}
\description{
Map between coin symbols and names
}
